render bug

+++ Simple-Raytracer/color.h
@@ -14,14 +14,14 @@ void write_color(std::fstream &file, color pixel_color, int samples_per_pixel) {
     auto b = pixel_color.z();
 
     /* Divide the color by the number of samples*/
-    auto scale = 1.0 / samples_per_pixel;
+    auto scale = 1.0 / static_cast<double>(samples_per_pixel);
     r = sqrt(scale * r);
     g = sqrt(scale * g);
     b = sqrt(scale * b);
 
-    file << static_cast<int>(256 * clamp(r, 0.0, .999)) << " "
-         << static_cast<int>(256 * clamp(g, 0.0, .999)) << " "
-         << static_cast<int>(256 * clamp(b, 0, .999)) << "\n";
+    file << static_cast<int>(static_cast<double>(256) * clamp(r, 0.0, .999)) << " "
+         << static_cast<int>(static_cast<double>(256) * clamp(g, 0.0, .999)) << " "
+         << static_cast<int>(static_cast<double>(256) * clamp(b, 0.0, .999)) << "\n";
 }
